UX Changes: ChatInput, ChatWindow 1. Move alerts to ChatInput
